Benchmark profiling for actions (#32)
Add a feature to make easy to measure the time spent for every action that goes through redukt. This is very useful to detect easily the bottlenecks. 

To enable the benchmark just set trace `true`

        redukt.traceActionProfile = true

This PR also solves a problem when new listeners are added during the dispatch, which raises a concurrent modification exception.